Move files out of UIAtoms directory
